+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.2.0](https://github.com/tbobm/dont-burn-it-all/compare/v0.1.0...v0.2.0) (2026-08-26)
+
+
+### Features
+
+* add opt-in --sandbox mode for isolated write sessions ([#24](https://github.com/tbobm/dont-burn-it-all/issues/24)) ([eadad0a](https://github.com/tbobm/dont-burn-it-all/commit/eadad0acf9559e5faf61160d7fdde8e6a55b50b2))
+
+## [0.1.0](https://github.com/tbobm/dont-burn-it-all/compare/v0.0.1...v0.1.0) (2026-08-05)
+
+
+### Features
+
+* **plugin:** self-host as marketplace for /plugin install ([#14](https://github.com/tbobm/dont-burn-it-all/issues/14)) ([f891752](https://github.com/tbobm/dont-burn-it-all/commit/f891752ff90f115673180769d4dd8e17b3ef7b0a))
+
 ## 0.0.1
 
 Initial release.